655-final-project-backend/cloud_functions_be_microservices/final-project-655-get-image-property-from-vision/requirements.txt
functions-framework==3.*
google-cloud-vision
requests
pillow
protobuf
jsonify